refactor: move onNodeDrag code to GraphEdit
